refactor: 프로필 이미지 설정 로직을 별도 컴포넌트로 분리
